requirements.txt
faster-whisper>=1.1.0
ctranslate2==4.4.0
anthropic==0.49.0
feedparser==6.0.11
flask==3.0.3
flask-cors==4.0.0
gunicorn==23.0.0
requests==2.32.3
python-slugify==8.0.4

# Audio analysis dependencies
librosa>=0.10.0
soundfile>=0.12.0
pyannote.audio>=3.1.0